Remove comment warning because Java is doing pass-by-ref like we want.
Extra mention: I saw Boese in her office hours about this use of the
State design pattern. She liked it so much that she offered us extra
credit if we send her the code for Table/TurnManager when we finish it
(she wants to use it as an example for her future classes). :)